Factored common parts out to common base class
